Move visitCallExpr to new file
